total_timeを0.1 s単位に、average_timeの計算を0.01 s単位に変更

+++ db/migrate/20200923051109_create_total_scores.rb
@@ -5,8 +5,8 @@ class CreateTotalScores < ActiveRecord::Migration[6.0]
       t.integer :total_correct_count, null: false
       t.integer :total_wrong_count, null: false
       t.integer :total_correct_rate, null: false
-      t.integer :total_time, null: false
-      t.integer :total_average_time, null: false
+      t.float :total_time, null: false
+      t.float :total_average_time, null: false
 
       t.timestamps
     end
